[morse_vision] Refactor and put morse vision into exe file
