+++ morse_vision/constants.py
