--- morse_vision/morse_functions.py
@@ -1,11 +1,0 @@
-def decode_morse_letter(letter):
-  morse_code = {
-    '.-': 'A', '-...': 'B', '-.-.': 'C', '-..': 'D', '.': 'E', '..-.': 'F', '--.': 'G',
-    '....': 'H', '..': 'I', '.---': 'J', '-.-': 'K', '.-..': 'L', '--': 'M', '-.': 'N',
-    '---': 'O', '.--.': 'P', '--.-': 'Q', '.-.': 'R', '...': 'S', '-': 'T', '..-': 'U',
-    '...-': 'V', '.--': 'W', '-..-': 'X', '-.--': 'Y', '--..': 'Z',
-  }
-  try:
-    return morse_code[letter.strip()]
-  except KeyError:
-    return 'Invalid morse letter'
